♻️ Move the remaining transfer tests into the core tests
